Mise à jour panier visiteur

+++ html/php/retrait_panier.php
@@ -3,48 +3,83 @@ require_once(__DIR__ . "/../01_premiere_connexion.php");
 
 session_start();
 
-// Récupération id du produit, id du client et le type de retrait
+// Récupération des infos par POST
 
-$idClient = $_SESSION["idCompte"];
 $idProd = $_POST["idProduit"];
 $typeRetrait = $_POST["typeRetrait"];
-$quantiteProd = $_POST["quantiteProd"];
-$quantiteTot = $_POST["quantiteTot"];
 $prixTTC = $_POST["prixTTC"];
-$prixTot = $_POST["prixTot"];
 
-if ($typeRetrait === "suppression" || $quantiteProd === "1") 
+
+if ($_SESSION['role']['client']) 
 {
-    // Suppression du produit du panier en supprimant la ligne correspondante dans _contient
+    // Récupération de l'id du client
+    $idClient = $_SESSION["idCompte"];
 
-    $rqt = $dbh->prepare("
-        DELETE FROM sae3_skadjam._contient
-        WHERE id_produit = :idProd
-    ");
+    if ($typeRetrait === "suppression" || $quantiteProd === "1") 
+    {
+        // Suppression du produit du panier en supprimant la ligne correspondante dans _contient
 
-    $rqt->execute([
-        ':idProd' => $idProd
-    ]);
+        $rqt = $dbh->prepare("
+            DELETE FROM sae3_skadjam._contient
+            WHERE id_produit = :idProd
+        ");
 
-    header("location:/html/fo/panier.php");
+        $rqt->execute([
+            ':idProd' => $idProd
+        ]);
+
+        header("location:/html/fo/panier.php");
+    }
+    else if ($typeRetrait === "decrement")
+    {
+        $quantiteProd--;
+
+        $rqt = $dbh->prepare("
+            UPDATE sae3_skadjam._contient SET
+            quantite_par_produit = :quantite
+            WHERE id_produit = :idProd
+        ");
+
+        $rqt->execute([
+            ':quantite' => $quantiteProd,
+            ':idProd' => $idProd
+        ]);
+
+        header("location:/html/fo/panier.php#" . $idProd);
+    }
 }
-else if ($typeRetrait === "decrement")
+else if ($_SESSION['role']['visiteur'])
 {
-    $quantiteProd--;
+    if ($typeRetrait === "suppression" || $quantiteProd === "1") 
+    {
+        // Suppression du produit du panier en supprimant la ligne correspondante dans _contient
 
-    $rqt = $dbh->prepare("
-        UPDATE sae3_skadjam._contient SET
-        quantite_par_produit = :quantite
-        WHERE id_produit = :idProd
-    ");
+        foreach ($_SESSION['panier']['contient'] as $i => $prod) 
+        {
+            if ($prod['id'] == $idProd) 
+            {
+                $_SESSION['montant_total_ttc'] -= $prixTTC;
+                $_SESSION['nb_produit_total'] -= $prod['quantite_par_produit'];
+                unset($_SESSION['panier']['contient'][$i]);
+                break;
+            }    
+        }
 
-    $rqt->execute([
-        ':quantite' => $quantiteProd,
-        ':idProd' => $idProd
-    ]);
+        header("location:/html/fo/panier.php");
+    }
+    else if ($typeRetrait === "decrement")
+    {
+        
+        
 
-    header("location:/html/fo/panier.php#" . $idProd);
+        header("location:/html/fo/panier.php#" . $idProd);
+    }
 }
+
+
+
+
+
